requirements.txt
django
gunicorn
whitenoise
psycopg2-binary
requests
urllib3
djangorestframework==3.15.2
djangorestframework-simplejwt
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.1
django-cors-headers
openai